+++ src/Fauna/reproduction_models.cpp
@@ -16,13 +16,25 @@ using namespace Fauna;
 // ##################################################################
 
 ReproductionLogistic::ReproductionLogistic(BreedingSeason breeding_season,
-                                           const double max_annual_increase)
+                                           const double max_annual_increase,
+                                           const double growth_rate,
+                                           const double midpoint)
     : max_annual_increase(max_annual_increase),
-      breeding_season(breeding_season) {
+      breeding_season(breeding_season),
+      growth_rate(growth_rate),
+      midpoint(midpoint) {
   if (max_annual_increase < 0.0)
     throw std::invalid_argument(
         "Fauna::ReproductionLogistic::ReproductionLogistic() "
         "max_annual_increase below zero.");
+  if (growth_rate <= 0.0)
+    throw std::invalid_argument(
+        "Fauna::ReproductionLogistic::ReproductionLogistic() "
+        "growth_rate is smaller than or equal to zero.");
+  if (midpoint <= 0.0 || midpoint >= 1.0)
+    throw std::invalid_argument(
+        "Fauna::ReproductionLogistic::ReproductionLogistic() "
+        "midpoint is not between zero and one.");
 }
 
 double ReproductionLogistic::get_offspring_density(
@@ -40,8 +52,8 @@ double ReproductionLogistic::get_offspring_density(
   if (!breeding_season.is_in_season(day_of_year)) return 0.0;
 
   // Yes, we are in breeding season and just apply the formula.
-  static const double b = 15.0;
-  static const double c = 0.3;
+  const double b = growth_rate;
+  const double c = midpoint;
   const double k = max_annual_increase;
 
   // see doxygen documentation of the class for formula